@@ -1,49 +1,60 @@
 {
   "name": "uploadbox",
-  "version": "1.0.0",
+  "version": "1.12.3",
   "description": "uploader like dropbox",
   "main": "./bundles/BrowseDrop.js",
   "author": "Marcos Josue Rangel Aguilar",
   "license": "MIT",
   "scripts": {
-    "start": "webpack-dev-server --inline --port 3000",
-    "build": "webpack --env=prod --progress --profile --colors --config=webpack.config.prod.js"
+    "startold": "webpack-dev-server --inline --port 3000 --mode development",
+    "buildold": "webpack --env=prod --progress --profile --colors --config=webpack.config.prod.js",
+    "start": "webpack-dev-server --open --mode development --progress",
+    "build": "webpack --mode production --progress --profile"
   },
   "peerDependencies": {
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2"
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1"
   },
   "dependencies": {
-    "animate.css": "^3.5.2",
-    "babel-core": "^6.18.2",
-    "babel-loader": "^6.2.7",
+    "@babel/polyfill": "^7.8.7",
+    "animate.css": "^3.7.2",
     "babel-polyfill": "^6.23.0",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
     "classnames": "^2.2.5",
     "deepmerge": "^1.3.2",
+    "e": "0.0.4",
     "json-loader": "^0.5.4",
-    "mime-types": "^2.1.15",
-    "prop-types": "^15.5.8",
-    "react": "^15.3.2",
-    "react-bootstrap": "^0.30.8",
-    "react-dom": "^15.3.2",
-    "react-icons": "^2.2.3",
-    "webpack": "^1.13.3",
+    "mime": "^3.0.0",
+    "prop-types": "^15.7.2",
+    "react": "^16.13.1",
+    "react-bootstrap": "^1.0.1",
+    "react-dom": "^16.13.1",
+    "react-icons": "^3.10.0",
     "whatwg-fetch": "^2.0.3"
   },
   "devDependencies": {
-    "babel-core": "^6.18.2",
+    "@babel/core": "^7.9.0",
+    "@babel/preset-env": "^7.9.5",
+    "@babel/preset-react": "^7.9.4",
+    "autoprefixer": "^9.7.6",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-loader": "^8.1.0",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
     "bootstrap": "^3.3.7",
     "compression-webpack-plugin": "^0.3.2",
-    "css-loader": "^0.26.2",
-    "style-loader": "^0.13.2",
-    "stylus": "^0.54.5",
-    "stylus-loader": "^2.5.1",
-    "webpack": "^1.14.0",
-    "webpack-dev-server": "^1.16.2"
+    "css-loader": "^6.7.3",
+    "html-loader": "^1.1.0",
+    "html-webpack-plugin": "^5.5.1",
+    "poststylus": "^1.0.0",
+    "rucksack-css": "^0.1.2",
+    "style-loader": "^1.2.0",
+    "stylus": "^0.54.7",
+    "stylus-loader": "^4.3.3",
+    "webpack": "^5.82.0",
+    "webpack-cli": "^5.0.2",
+    "webpack-dev-server": "^4.13.3"
   },
   "jshintConfig": {
     "esversion": 6
